src/server/requirements.txt
pyaudio
pynput
pyserial